account_service/requirements.txt
fastapi==0.70.0
alembic==1.7.5
uvicorn==0.15.0
SQLAlchemy==1.4.27
orjson==3.6.4
alembic==1.7.5
pyjwt==2.3.0
httpx==0.21.1
gunicorn==20.1.0
sqlalchemy-utils==0.37.9
python-dotenv==0.19.2
faker==9.9.0
fastapi-pagination==0.9.1
requests==2.26.0
asgi_lifespan==1.0.1
asyncpg==0.25.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==1.9.0